Refactor: Move matrix operations to src/utils/matrix_operations
