+++ src/utils/feature_selection/step08_unified_complete.py
@@ -49,8 +49,8 @@ try:
     from src.utils.hardware.m1_gpu_utils import get_m1_gpu_manager, M1GPUManager
     from src.utils.hardware.m1_memory_optimizer import get_m1_memory_optimizer, M1MemoryOptimizer
     from src.utils.hardware.m1_cpu_optimizer import get_m1_cpu_optimizer, M1CPUOptimizer
-    from src.utils.vectorized_processing_core import OptimizedPipelineExecutor, PipelineStage, PipelineExecutionMode
-    from src.utils.ml_common.matrix_operations import EnhancedMatrixOperations, ErrorHandler
+    from src.utils.matrix_operations import OptimizedPipelineExecutor, PipelineStage, PipelineExecutionMode
+    from src.utils.matrix_operations import EnhancedMatrixOperations, ErrorHandler
     from src.utils.enhanced_step_optimizations import IntelligentOptimizationSelector, OptimizationStrategy, WorkloadType, OptimizationProfile
     from src.utils.optimized_data_manager import OptimizedDataManager, DataMetadata
     ENHANCED_OPTIMIZATIONS_AVAILABLE = True
